--- a/controller/BootstrapController.go
+++ b/controller/BootstrapController.go
@@ -7,16 +7,18 @@ import (
 
 	"github.com/CalebRose/SimFBA/managers"
 	"github.com/gorilla/mux"
+	"github.com/mailru/easyjson"
 )
 
 func BootstrapTeamData(w http.ResponseWriter, r *http.Request) {
 	data := managers.GetTeamsBootstrap()
 	w.Header().Set("Content-Type", "application/json")
-	err := json.NewEncoder(w).Encode(data)
+	teamData, err := easyjson.Marshal(data)
 	if err != nil {
 		log.Printf("Failed to encode JSON response: %v", err)
 		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
 	}
+	w.Write(teamData)
 }
 
 func FirstBootstrapFootballData(w http.ResponseWriter, r *http.Request) {
@@ -25,7 +27,12 @@ func FirstBootstrapFootballData(w http.ResponseWriter, r *http.Request) {
 	collegeID := vars["collegeID"]
 	proID := vars["proID"]
 	data := managers.GetFirstBootstrapData(collegeID, proID)
-	json.NewEncoder(w).Encode(data)
+	bootstrapData, err := easyjson.Marshal(data)
+	if err != nil {
+		log.Printf("Failed to encode JSON response: %v", err)
+		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+	}
+	w.Write(bootstrapData)
 }
 
 func SecondBootstrapFootballData(w http.ResponseWriter, r *http.Request) {
@@ -34,7 +41,12 @@ func SecondBootstrapFootballData(w http.ResponseWriter, r *http.Request) {
 	collegeID := vars["collegeID"]
 	proID := vars["proID"]
 	data := managers.GetSecondBootstrapData(collegeID, proID)
-	json.NewEncoder(w).Encode(data)
+	bootstrapData, err := easyjson.Marshal(data)
+	if err != nil {
+		log.Printf("Failed to encode JSON response: %v", err)
+		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+	}
+	w.Write(bootstrapData)
 }
 
 func ThirdBootstrapFootballData(w http.ResponseWriter, r *http.Request) {
@@ -43,7 +55,12 @@ func ThirdBootstrapFootballData(w http.ResponseWriter, r *http.Request) {
 	collegeID := vars["collegeID"]
 	proID := vars["proID"]
 	data := managers.GetThirdBootstrapData(collegeID, proID)
-	json.NewEncoder(w).Encode(data)
+	bootstrapData, err := easyjson.Marshal(data)
+	if err != nil {
+		log.Printf("Failed to encode JSON response: %v", err)
+		http.Error(w, "Internal Server Error", http.StatusInternalServerError)
+	}
+	w.Write(bootstrapData)
 }
 
 func GetCollegeHistoryProfile(w http.ResponseWriter, r *http.Request) {