--- a/go.mod
+++ b/go.mod
@@ -15,7 +15,11 @@ require (
 	gorm.io/gorm v1.25.5
 )
 
-require github.com/tkrajina/go-reflector v0.5.5 // indirect
+require (
+	github.com/josharian/intern v1.0.0 // indirect
+	github.com/mailru/easyjson v0.9.0 // indirect
+	github.com/tkrajina/go-reflector v0.5.5 // indirect
+)
 
 require (
 	github.com/felixge/httpsnoop v1.0.4 // indirect