Attempts to make several performance improvements on the API. From testing, the Facedata retrieval was the longest part of the Bootstrap 1 call, and shouldn't be necessary to displaying the initial homepage. This has been moved to Bootstrap 3 to let that finish in the background. Additionally, for the football bootstrap data, EasyJSON a new, more performant JSON encoder, was employed to allow the API to marshal the JSON responses to the bootstap calls more quickly.

Based on local testing, the response time for Bootstrap 1 has been improved by about 2 seconds, and although Bootstrap 3 takes about a second longer now, it happens in the background after the page loads, so it should be less noticeable to the user. Response times in Azure seems to be significantly different compared to local, so the final impact there probably won't be known until merge, but it should be faster.

NOTE: This will require a small change on the V2 interface to work correctly. The PR for this change is [here](https://github.com/CalebRose/simsn-interface-v2/pull/94). If either change is deployed without the other, attempting to open a player card will crash the page!

Local testing results:
Main branch, without changes:
![image](https://github.com/user-attachments/assets/e1b86423-1c9a-47cd-a226-cb867ff2b3ba)
Bootstrap one response time: 5148ms
Bootstrap three response time: 4210ms

Performance branch, with changes:
![image](https://github.com/user-attachments/assets/91ef3ef4-ba28-445a-bf5c-e9ce2c1a7edf)
Bootstrap one response time: 3117ms
Bootstrap three response time: 5351ms